requirements.txt
requests
telebot
pillow